+++ test/index-test.js
@@ -3,35 +3,42 @@ const _ = require('lodash')
 const provider = require('./fixtures/fake-provider')
 const auth = require('./fixtures/fake-auth')()
 const Koop = require('../src/')
-const Controller = require('../src/controllers')
+const Geoservices = require('koop-output-geoservices')
 const request = require('supertest')
 const should = require('should') // eslint-disable-line
 const plugins = require('./fixtures/fake-plugin')
+const geoservicesFixtureRoutes = Geoservices.routes.reduce((acc, route) => {
+  return acc + route.methods.length
+}, 0)
+const providerFixtureRoutes = provider.routes.reduce((acc, route) => {
+  return acc + route.methods.length
+}, 0)
 
-describe('Index tests for registering providers', function () {
-  describe('use config argument', function () {
-    it('should register successfully', function () {
+describe('Index tests', function () {
+  describe('Koop instantiation', function () {
+    it('should instantiate Koop with config', function () {
       const koop = new Koop({ foo: 'bar' })
       koop.config.should.have.property('foo', 'bar')
     })
   })
 
-  describe('provider registration', function () {
-    it('should register successfully', function () {
+  describe('Provider registration', function () {
+    it('should register provider and add output and provider routes to router stack', function () {
       const koop = new Koop()
-      koop.register(_.cloneDeep(provider))
-      koop.controllers.should.have.property('test-provider').and.be.a.instanceOf(Controller)
+      koop.register(provider)
+      const registeredProvider = koop.providers.find(provider => { return provider.namespace === 'test-provider' })
+      registeredProvider.should.have.property('namespace', 'test-provider')
       // Check that the stack includes routes with the provider name in the path
-      const providerPath = koop.server._router.stack
+      const providerRoutes = koop.server._router.stack
         .filter((layer) => { return _.has(layer, 'route.path') })
         .map(layer => { return _.get(layer, 'route.path') })
-        .find(path => path.includes(provider.name))
-      providerPath.should.not.equal(undefined)
+        .filter(path => path.includes(provider.name) || path.includes('/fake/:id'))
+      providerRoutes.length.should.equal(geoservicesFixtureRoutes + providerFixtureRoutes)
     })
 
     it('should register provider-routes before plugin-routes', function () {
       const koop = new Koop()
-      koop.register(_.cloneDeep(provider))
+      koop.register(provider)
       // Check that the stack index of the plugin routes are prior to index of provider routes
       const routePaths = koop.server._router.stack
         .filter((layer) => { return _.has(layer, 'route.path') })
@@ -43,7 +50,7 @@ describe('Index tests for registering providers', function () {
 
     it('should register plugin-routes before provider-routes', function () {
       const koop = new Koop()
-      koop.register(_.cloneDeep(provider), { defaultToOutputRoutes: true })
+      koop.register(provider, { defaultToOutputRoutes: true })
       // Check that the stack index of the plugin routes are prior to index of provider routes
       const routePaths = koop.server._router.stack
         .filter((layer) => { return _.has(layer, 'route.path') })
@@ -53,31 +60,40 @@ describe('Index tests for registering providers', function () {
       pluginRouteIndex.should.be.below(providerRouteIndex)
     })
 
-    it('should register successfully and attach cache and options object to model', function () {
+    it('should register successfully a provider with a routePrefix', function () {
       const koop = new Koop()
-      koop.register(_.cloneDeep(provider), { routePrefix: 'path-to-route' })
-      koop.controllers['test-provider'].model.should.have.property('cache')
-      koop.controllers['test-provider'].model.should.have.property('options')
-      koop.controllers['test-provider'].model.options.should.have.property('routePrefix', 'path-to-route')
+      koop.register(provider, { routePrefix: 'path-to-route' })
+      const registeredProvider = koop.providers.find(provider => { return provider.namespace === 'test-provider' })
+      registeredProvider.options.should.have.property('routePrefix', 'path-to-route')
     })
 
-    it('should register successfully and attach optional cache to model', function () {
+    it('should register successfully and attach cache and options object to model', function () {
       const koop = new Koop()
-      koop.register(_.cloneDeep(provider), {
+      koop.register(provider, { foo: 'bar' })
+      const registeredProvider = koop.providers.find(provider => { return provider.namespace === 'test-provider' })
+      registeredProvider.model.should.have.property('cache')
+      registeredProvider.model.should.have.property('options')
+      registeredProvider.model.options.should.have.property('foo', 'bar')
+    })
+
+    it('should register successfully and attach optional custom cache to model', function () {
+      const koop = new Koop()
+      koop.register(provider, {
         cache: {
           retrieve: (key, query, callback) => {},
           upsert: (key, data, options) => {},
           customFunction: () => {}
         }
       })
-      koop.controllers['test-provider'].model.should.have.property('cache')
-      koop.controllers['test-provider'].model.cache.should.have.property('customFunction').and.be.a.Function()
+      const registeredProvider = koop.providers.find(provider => { return provider.namespace === 'test-provider' })
+      registeredProvider.model.should.have.property('cache')
+      registeredProvider.model.cache.should.have.property('customFunction').and.be.a.Function()
     })
 
     it('should reject cache option missing an upsert method', function () {
       const koop = new Koop()
       try {
-        koop.register(_.cloneDeep(provider), {
+        koop.register(provider, {
           cache: {
             retrieve: (key, query, callback) => {}
           }
@@ -90,7 +106,7 @@ describe('Index tests for registering providers', function () {
     it('should reject cache option missing a retrieve method', function () {
       const koop = new Koop()
       try {
-        koop.register(_.cloneDeep(provider), {
+        koop.register(provider, {
           cache: {
             upsert: (key, data, options) => {}
           }
@@ -103,7 +119,7 @@ describe('Index tests for registering providers', function () {
     it('should reject routePrefix option that is not a string', function () {
       const koop = new Koop()
       try {
-        koop.register(_.cloneDeep(provider), {
+        koop.register(provider, {
           routePrefix: {}
         })
         should.fail('should have thrown error')
@@ -114,18 +130,19 @@ describe('Index tests for registering providers', function () {
 
     it('should register successfully and attach optional "before" and "after" function to model', function () {
       const koop = new Koop()
-      koop.register(_.cloneDeep(provider), {
+      koop.register(provider, {
         before: (req, next) => {},
         after: (req, data, callback) => {}
       })
-      koop.controllers['test-provider'].model.should.have.property('before').and.be.a.Function()
-      koop.controllers['test-provider'].model.should.have.property('after').and.be.a.Function()
+      const registeredProvider = koop.providers.find(provider => { return provider.namespace === 'test-provider' })
+      registeredProvider.model.should.have.property('before').and.be.a.Function()
+      registeredProvider.model.should.have.property('after').and.be.a.Function()
     })
 
     it('should reject optional "before" function that does not have correct arity', function () {
       const koop = new Koop()
       try {
-        koop.register(_.cloneDeep(provider), {
+        koop.register(provider, {
           before: () => {}
         })
       } catch (err) {
@@ -136,7 +153,7 @@ describe('Index tests for registering providers', function () {
     it('should reject optional "after" function that does not have correct arity', function () {
       const koop = new Koop()
       try {
-        koop.register(_.cloneDeep(provider), {
+        koop.register(provider, {
           after: () => {}
         })
       } catch (err) {
@@ -146,24 +163,24 @@ describe('Index tests for registering providers', function () {
 
     it('should successfully use options "name" in route', function () {
       const koop = new Koop()
-      koop.register(_.cloneDeep(provider), {
+      koop.register(provider, {
         name: 'options-name'
       })
-      koop.controllers.should.have.property('options-name').and.be.a.instanceOf(Controller)
-      koop.controllers.should.have.property('options-name').and.be.a.instanceOf(Controller)
+      const registeredProvider = koop.providers.find(provider => { return provider.namespace === 'options-name' })
+      registeredProvider.should.have.property('namespace', 'options-name')
       // Check that the stack includes routes with the provider name in the path
-      const providerPath = koop.server._router.stack
+      const providerRoutes = koop.server._router.stack
         .filter((layer) => { return _.has(layer, 'route.path') })
         .map(layer => { return _.get(layer, 'route.path') })
-        .find(path => path.includes('options-name'))
-      providerPath.should.not.equal(undefined)
+        .filter(path => path.includes('options-name'))
+      providerRoutes.length.should.equal(geoservicesFixtureRoutes)
     })
   })
 
   describe('can register a provider and apply a route prefix to all routes', function () {
     it('should not return 404 for prefixed custom route', function (done) {
       const koop = new Koop()
-      koop.register(_.cloneDeep(provider), { routePrefix: '/api/test' })
+      koop.register(provider, { routePrefix: '/api/test' })
       request(koop.server)
         .get('/api/test/fake/1234')
         .then((res) => {
@@ -177,7 +194,7 @@ describe('Index tests for registering providers', function () {
 
     it('should not return 404 for prefixed plugin route', function (done) {
       const koop = new Koop()
-      koop.register(_.cloneDeep(provider), { routePrefix: '/api/test' })
+      koop.register(provider, { routePrefix: '/api/test' })
       request(koop.server)
         .get('/api/test/test-provider/foo/FeatureServer')
         .then((res) => {