+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* Refactored provider registration, output registration, and controller treatment.
+
 ## [3.18.0] - 2020-03-31
 ### Added
 * Option to allow output routes to be registered before provider routes; if true any conflicts will default to output routes